refactor to use subdir and change some names
