+++ scipy/integrate/_rules/__init__.py
@@ -1,0 +1,7 @@
+# flake8: noqa
+
+from ._base import *
+from ._genz_malik import GenzMalikCub
+from ._newton_cotes import NewtonCotesQuad
+from ._gauss_kronrod import GaussKronrodQuad
+from ._gauss_legendre import GaussLegendreQuad
