+++ scipy/integrate/_rules/meson.build
@@ -1,0 +1,10 @@
+py3.install_sources([
+    '__init__.py',
+    '_base.py',
+    '_gauss_kronrod.py',
+    '_gauss_legendre.py',
+    '_genz_malik.py',
+    '_newton_cotes.py',
+  ],
+  subdir: 'scipy/integrate/_rules'
+)
